Renamed file to init and modified structure to have it in a folder to then zip fo Blender
